v 0.02 commit: crud for users and comment section done. fix bugs with editing and deleting not your posts

+++ myblog/feed/views.py
@@ -1,6 +1,7 @@
 import datetime
 
 from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 from django.shortcuts import get_object_or_404, redirect, render
 
 import comments.forms
@@ -65,30 +66,38 @@ def create_post(request):
 def edit_post(request, post_id):
     template = "feed/includes/edit_post.html"
     post = get_object_or_404(feed.models.Feed, pk=post_id)
-    if request.method == "POST":
-        form = feed.forms.CreatePostFrom(request.POST,
-                                         request.FILES,
-                                         instance=post)
-        if form.is_valid():
-            post = form.save(commit=False)
-            post.date_edited = datetime.datetime.now()
-            post.save()
-            return redirect("items_list")
+    user = get_object_or_404(feed.models.users.models.User, id=post.user.id)
+    if request.user.id == user.id:
+        if request.method == "POST":
+            form = feed.forms.CreatePostFrom(request.POST,
+                                             request.FILES,
+                                             instance=post)
+            if form.is_valid():
+                post = form.save(commit=False)
+                post.date_edited = datetime.datetime.now()
+                post.save()
+                return redirect("items_list")
+        else:
+            form = feed.forms.CreatePostFrom(instance=post)
     else:
-        form = feed.forms.CreatePostFrom(instance=post)
+        return HttpResponse("<h1>You can't edit not your posts!</h1>")
     context = {
         "post": post,
         "form": form,
     }
-
+    
     return render(request, template, context)
 
 
 @login_required
 def delete_post(request, post_id):
-    if request.method == "POST":
+    post = get_object_or_404(feed.models.Feed, id=post_id)
+    user = get_object_or_404(feed.models.users.models.User, id=post.user.id)
+    if request.user.id == user.id and request.method == "POST":
         post = get_object_or_404(feed.models.Feed, pk=post_id)
         post.delete()
+    else:
+        return HttpResponse("<h1>You can't delete not your posts!</h1>")
     return redirect("items_list")
 
 
@@ -119,38 +128,44 @@ def dislike(request, post_id):
 
 
 @login_required
-def edit_comment(request, comment_id):
+def edit_comment(request, post_id, comment_id):
     template = "feed/includes/edit_comment.html"
     comment = get_object_or_404(comments.models.Comments, pk=comment_id)
-    if request.method == "POST":
-        form = comments.forms.EditCommentForm(request.POST, instance=comment)
-        if form.is_valid():
-            form.save(commit=False)
-            return redirect("items_list")
+    user = get_object_or_404(feed.models.users.models.User, id=comment.user.id)
+    if request.user.id == user.id:
+        if request.method == "POST":
+            form = comments.forms.EditCommentForm(request.POST, request.FILES, instance=comment)
+            if form.is_valid():
+                form.save()
+                return redirect("item_detail", post_id)
+        else:
+            form = comments.forms.EditCommentForm(instance=comment)
+        context = {
+            "comment": comment,
+            "form": form,
+        }
     else:
-        form = feed.forms.CreatePostFrom(instance=comment)
-    context = {
-        "comment": comment,
-        "form": form,
-    }
-
+        return HttpResponse("<h1>You can't edit not your comments!</h1>")
     return render(request, template, context)
 
 
 @login_required
 def delete_comment(request, comment_id):
     comment = get_object_or_404(comments.models.Comments, id=comment_id)
-    if request.method == "POST" and comment.user == request.user:
-        comment.delete()
+    user = get_object_or_404(feed.models.users.models.User, id=comment.user.id)
+    if request.user.id == user.id:
+        if request.method == "POST" and comment.user == request.user:
+            comment.delete()
+    else:
+        return HttpResponse("<h1>You can't delete not your comments!</h1>")
     return redirect("items_list")
 
 
 def switch_theme(request):
     current_theme = request.session.get("theme", "light")
-    # Toggle theme
     if current_theme == "light":
         request.session["theme"] = "dark"
     else:
         request.session["theme"] = "light"
-    # Redirect to the previous page or homepage
+        
     return redirect(request.META.get("HTTP_REFERER", "/"))
